# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MSS Sample Stocking Platform API
  version: 1.0.0
  description: Contract for overview, demand collection, execution tracking and master data configuration.
servers:
  - url: http://localhost:8787/api/v1
...
security:
  - bearerAuth: []
paths:
...
  /execution/imports:
    get:
      tags: [Execution]
      summary: List the current stocking owner's recent TSMP import jobs
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Execution]
      summary: Create a TSMP shipment import job
      parameters:
        - $ref: '#/components/parameters/RoleHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fileName]
              properties:
                fileName: { type: string }
                rows:
                  type: array
                  description: Local starter accepts parsed rows; production accepts a controlled upload reference.
                  items:
                    $ref: '#/components/schemas/TsmpShipmentRow'
      responses:
        '202': { $ref: '#/components/responses/Success' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
  /execution:
    get:
      tags: [Execution]
      summary: Get product-first cumulative execution tracking
      parameters:
        - $ref: '#/components/parameters/ProductId'
        - $ref: '#/components/parameters/RegionId'
        - name: officeId
          in: query
          schema: { type: string }
        - name: country
          in: query
          schema: { type: string }
        - name: keyword
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Metrics and product rows with nested SKUs
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/ExecutionView' }
...
  /config/products:
    post:
      tags: [Configuration]
      summary: Create a product; SKU and BOM are optional
      parameters:
        - $ref: '#/components/parameters/RoleHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProductInput' }
      responses:
        '201': { $ref: '#/components/responses/Success' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /config/products/{productId}:
    put:
      tags: [Configuration]
      summary: Update product configuration
      parameters:
        - $ref: '#/components/parameters/ProductIdPath'
        - $ref: '#/components/parameters/RoleHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProductInput' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '404': { $ref: '#/components/responses/NotFound' }
  /config/domains:
    post:
      tags: [Configuration]
      summary: Create a product domain responsibility record
      parameters:
        - $ref: '#/components/parameters/RoleHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DomainInput' }
      responses:
        '201': { $ref: '#/components/responses/Success' }
  /config/domains/{domainId}:
    put:
      tags: [Configuration]
      summary: Update domain and inherited owners
      parameters:
        - name: domainId
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/RoleHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DomainInput' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /config/organizations:
    post:
      tags: [Configuration]
      summary: Create a region with representative offices
      parameters:
        - $ref: '#/components/parameters/RoleHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OrganizationInput' }
      responses:
        '201': { $ref: '#/components/responses/Success' }
  /config/organizations/{regionId}:
    put:
      tags: [Configuration]
      summary: Update a region tree
      parameters:
        - $ref: '#/components/parameters/RegionIdPath'
        - $ref: '#/components/parameters/RoleHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OrganizationInput' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    RoleHeader:
      name: Authorization
      in: header
      required: true
      description: JWT bearer token. Retained as a reusable parameter for existing operation entries.
      schema: { type: string, example: 'Bearer eyJ...' }
...
    ProductId:
      name: productId
      in: query
      schema: { type: string, default: all }
    ProductIdPath:
      name: productId
      in: path
      required: true
      schema: { type: string }
    RegionId:
      name: regionId
      in: query
      schema: { type: string }
    RegionIdPath:
      name: regionId
      in: path
      required: true
      schema: { type: string }
  responses:
    Success:
      description: Successful response
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
    Forbidden:
      description: Role or scope denied
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
    Envelope:
      type: object
      required: [code, message, requestId]
      properties:
        code: { type: string, example: OK }
        message: { type: string, example: success }
        data: {}
        requestId: { type: string }
    ErrorEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            details:
              type: array
              items: { type: string }
...
    ProductInput:
      type: object
      required: [name, domainId]
      properties:
        name: { type: string, minLength: 1 }
        domainId: { type: string }
        mssDomainId: { type: string }
        stage: { type: string }
        supplyTimeText: { type: string }
        defaultDeadline: { type: string }
        defaultRegionIds: { type: array, items: { type: string } }
        enabled: { type: boolean, default: true }
        skus:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              model: { type: string }
              bomCode: { type: string }
              description: { type: string }
    DomainInput:
      type: object
      required: [name, gtmOwner, stockingOwner]
      properties:
        name: { type: string }
        description: { type: string }
        gtmOwner: { type: string }
        stockingOwner: { type: string }
        enabled: { type: boolean, default: true }
    OrganizationInput:
      type: object
      required: [name, owner]
      properties:
        name: { type: string }
        owner: { type: string }
        enabled: { type: boolean, default: true }
        offices:
          type: array
          items:
            type: object
            required: [name, owner]
            properties:
              id: { type: string }
              name: { type: string }
              owner: { type: string }
              enabled: { type: boolean, default: true }
              countries: { type: array, items: { type: string } }
    TsmpShipmentRow:
      type: object
      required: [externalKey, sku, region, office, shippedQty]
      properties:
        externalKey: { type: string }
        applicationNo: { type: string }
        sku: { type: string }
        bomCode: { type: string }
        region: { type: string }
        office: { type: string }
        country: { type: string }
        shippedQty: { type: integer, minimum: 1 }
        shippedAt: { type: string, format: date-time }
...
    ExecutionSku:
      type: object
      properties:
        sku: { type: string }
        bom: { type: string }
        demand: { type: integer }
        applied: { type: integer }
        shipped: { type: integer }
        inventory: { type: integer }
        shipmentCount: { type: integer }
    ExecutionView:
      type: object
      properties:
        scopeLabel: { type: string }
        metrics: { type: object, additionalProperties: { type: integer } }
        products:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              name: { type: string }
              domain: { type: string }
              gtm: { type: string }
              stockingOwner: { type: string }
              metrics: { type: object, additionalProperties: { type: integer } }
              skus:
                type: array
                items: { $ref: '#/components/schemas/ExecutionSku' }
...